+++ app/uk/gov/hmrc/tai/model/domain/income/TaxCodeIncome.scala
@@ -41,29 +41,6 @@ object BasisOfOperation {
   }
 }
 
-// TODO:// Move to employment model or own file
-sealed trait TaxCodeIncomeSourceStatus
-case object Live extends TaxCodeIncomeSourceStatus
-case object NotLive extends TaxCodeIncomeSourceStatus
-case object PotentiallyCeased extends TaxCodeIncomeSourceStatus
-case object Ceased extends TaxCodeIncomeSourceStatus
-
-object TaxCodeIncomeSourceStatus {
-  implicit val formatTaxCodeIncomeSourceStatus: Format[TaxCodeIncomeSourceStatus] =
-    new Format[TaxCodeIncomeSourceStatus] {
-      override def reads(json: JsValue): JsResult[TaxCodeIncomeSourceStatus] = json.as[String] match {
-        case "Live"              => JsSuccess(Live)
-        case "NotLive"           => JsSuccess(NotLive)
-        case "PotentiallyCeased" => JsSuccess(PotentiallyCeased)
-        case "Ceased"            => JsSuccess(Ceased)
-        case _                   => JsError("Invalid Tax component type")
-      }
-
-      override def writes(taxCodeIncomeSourceStatus: TaxCodeIncomeSourceStatus) =
-        JsString(taxCodeIncomeSourceStatus.toString)
-    }
-}
-
 sealed trait IabdUpdateSource
 
 case object ManualTelephone extends IabdUpdateSource
