+++ test/uk/gov/hmrc/tai/model/domain/CodingComponentFormattersSpec.scala
@@ -17,7 +17,7 @@
 package uk.gov.hmrc.tai.model.domain
 
 import org.scalatestplus.play.PlaySpec
-import play.api.libs.json.{JsError, JsResultException, JsString, Json}
+import play.api.libs.json.{JsString, Json}
 import uk.gov.hmrc.tai.model.domain.calculation.CodingComponent
 import uk.gov.hmrc.tai.model.domain.formatters.CodingComponentFormatters
 import uk.gov.hmrc.tai.model.domain.income._
@@ -237,29 +237,6 @@ class CodingComponentFormattersSpec extends PlaySpec with CodingComponentFormatt
           "employment2",
           Week1Month1BasisOfOperation,
           Ceased)
-      }
-    }
-  }
-
-  // TODO:// move this
-  "taxCodeIncomeSourceStatusReads" must {
-    "read the field correctly" when {
-      "json string is Live" in {
-        JsString("Live").as[TaxCodeIncomeSourceStatus] mustBe Live
-      }
-      "json string is PotentiallyCeased" in {
-        JsString("PotentiallyCeased")
-          .as[TaxCodeIncomeSourceStatus] mustBe PotentiallyCeased
-      }
-      "json string is Ceased" in {
-        JsString("Ceased").as[TaxCodeIncomeSourceStatus] mustBe Ceased
-      }
-    }
-    "throw JsResultException" when {
-      "provided with unrecognized status" in {
-        val ex = the[JsResultException] thrownBy JsString("Some Status")
-          .as[TaxCodeIncomeSourceStatus]
-        ex.getMessage must include("Invalid Tax component type")
       }
     }
   }
